requirements.txt
django>=1.4,<1.7
Markdown>2.2.0
django-sekizai>=0.7
django-configurations==0.8
dj-database-url==0.3.0
django-cache-url==0.8.0
psycopg2==2.5

# django-mptt 0.6.1 is broken, see #316
django-mptt==0.6.0

# Now with compatibility for django 1.6
sorl-thumbnail>=11.12.1b

Pillow
six
# Notification system
django-nyt>=0.9.3
